requirements.txt
Flask==3.0.3
Flask-Limiter==3.8.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
SQLAlchemy==2.0.32
Werkzeug==3.0.3
PyYAML==6.0.2
python-dotenv==1.0.1
reportlab==4.2.2
requests==2.32.3
pytest==8.3.4
